# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified API for GambiaInfluencers and Our Influencers platforms
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Influencer {
  id: number;
  name: string;
  location: string;
  niche: string;
  followersCount: number;
  /** @nullable */
  instagramUrl: string | null;
  /** @nullable */
  tiktokUrl: string | null;
  /** @nullable */
  youtubeUrl: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  whatsappNumber: string | null;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateInfluencerBody {
  name: string;
  location: string;
  niche: string;
  followersCount: number;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  tiktokUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
}

export interface UpdateInfluencerBody {
  name?: string;
  location?: string;
  niche?: string;
  followersCount?: number;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  tiktokUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
}

export interface Business {
  id: number;
  businessName: string;
  businessType: string;
  /** @nullable */
  contactEmail: string | null;
  /** @nullable */
  contactPhone: string | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  website: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBusinessBody {
  businessName: string;
  businessType: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  website?: string | null;
}

export interface UpdateBusinessBody {
  businessName?: string;
  businessType?: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  website?: string | null;
}

export interface PlatformStats {
  totalInfluencers: number;
  totalBusinesses: number;
  totalFollowers: number;
  topNiche: string;
  topLocation: string;
}

export interface NicheCount {
  niche: string;
  count: number;
}

export interface LocationCount {
  location: string;
  count: number;
}

export interface OiLoginBody {
  email: string;
  password: string;
}

export type OiUserRole = (typeof OiUserRole)[keyof typeof OiUserRole];

export const OiUserRole = {
  admin: "admin",
  client: "client",
} as const;

export interface OiUser {
  id: number;
  email: string;
  fullName: string;
  role: OiUserRole;
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  phone: string | null;
  createdAt: string;
}

export interface OiAuthResponse {
  token: string;
  user: OiUser;
}

export type OiCreateUserBodyRole =
  (typeof OiCreateUserBodyRole)[keyof typeof OiCreateUserBodyRole];

export const OiCreateUserBodyRole = {
  admin: "admin",
  client: "client",
} as const;

export interface OiCreateUserBody {
  email: string;
  password: string;
  fullName: string;
  role: OiCreateUserBodyRole;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface OiUpdateUserBody {
  email?: string;
  fullName?: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  password?: string | null;
}

export type OiCampaignStatus =
  (typeof OiCampaignStatus)[keyof typeof OiCampaignStatus];

export const OiCampaignStatus = {
  pending: "pending",
  active: "active",
  posted: "posted",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface OiCampaign {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  budget?: string | null;
  status: OiCampaignStatus;
  /** @nullable */
  postUrl?: string | null;
  /** @nullable */
  postedAt?: string | null;
  clientId: number;
  influencerId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  influencerName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type OiCampaignDetailStatus =
  (typeof OiCampaignDetailStatus)[keyof typeof OiCampaignDetailStatus];

export const OiCampaignDetailStatus = {
  pending: "pending",
  active: "active",
  posted: "posted",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface OiCampaignDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  budget?: string | null;
  status: OiCampaignDetailStatus;
  /** @nullable */
  postUrl?: string | null;
  /** @nullable */
  postedAt?: string | null;
  clientId: number;
  influencerId: number;
  client?: OiUser;
  influencer?: Influencer;
  createdAt: string;
  updatedAt: string;
}

export interface OiCreateCampaignBody {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  budget?: string | null;
  clientId: number;
  influencerId: number;
}

export type OiUpdateCampaignBodyStatus =
  (typeof OiUpdateCampaignBodyStatus)[keyof typeof OiUpdateCampaignBodyStatus];

export const OiUpdateCampaignBodyStatus = {
  pending: "pending",
  active: "active",
  posted: "posted",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface OiUpdateCampaignBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  budget?: string | null;
  status?: OiUpdateCampaignBodyStatus;
  /** @nullable */
  postUrl?: string | null;
}

export interface OiCampaignSummary {
  total: number;
  pending: number;
  active: number;
  posted: number;
  completed: number;
  cancelled: number;
  totalBudget: string;
}

export type OiQuickPromotionStatus =
  (typeof OiQuickPromotionStatus)[keyof typeof OiQuickPromotionStatus];

export const OiQuickPromotionStatus = {
  new: "new",
  contacted: "contacted",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface OiQuickPromotion {
  id: number;
  influencerId: number;
  /** @nullable */
  influencerName?: string | null;
  contactName: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  description: string;
  promoType: string;
  status: OiQuickPromotionStatus;
  createdAt: string;
}

export interface OiCreateQuickPromotionBody {
  influencerId: number;
  contactName: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  description: string;
  promoType: string;
}

export type OiUpdateQuickPromotionBodyStatus =
  (typeof OiUpdateQuickPromotionBodyStatus)[keyof typeof OiUpdateQuickPromotionBodyStatus];

export const OiUpdateQuickPromotionBodyStatus = {
  new: "new",
  contacted: "contacted",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface OiUpdateQuickPromotionBody {
  status?: OiUpdateQuickPromotionBodyStatus;
}

export type ListInfluencersParams = {
  location?: string;
  niche?: string;
  search?: string;
};

export type GetTopInfluencersParams = {
  limit?: number;
};