# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Unified API for GambiaInfluencers and Our Influencers platforms
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /oi/auth/login:
    post:
      operationId: oiLogin
      tags: [auth]
      summary: Login as admin or client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OiLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OiAuthResponse"
        "401":
          description: Invalid credentials
  /oi/auth/me:
    get:
      operationId: oiGetMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OiUser"
        "401":
          description: Unauthorized

  # ─── Our Influencers: Users ──────────────────────────────────────────────────
...
  /oi/campaigns:
    get:
      operationId: oiListCampaigns
      tags: [campaigns]
      summary: List campaigns (admin sees all, client sees own)
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OiCampaign"
    post:
      operationId: oiCreateCampaign
      tags: [campaigns]
      summary: Create campaign (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OiCreateCampaignBody"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OiCampaign"
  /oi/campaigns/{id}:
    get:
      operationId: oiGetCampaign
      tags: [campaigns]
      summary: Get campaign details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OiCampaignDetail"
        "404":
          description: Not found
    patch:
      operationId: oiUpdateCampaign
      tags: [campaigns]
      summary: Update campaign status or post URL (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OiUpdateCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OiCampaign"
    delete:
      operationId: oiDeleteCampaign
      tags: [campaigns]
      summary: Delete campaign (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /oi/quick-promotions:
    get:
      operationId: oiListQuickPromotions
      tags: [quickPromotions]
      summary: List quick promotion requests (admin only)
      responses:
        "200":
          description: List of quick promotions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OiQuickPromotion"
    post:
      operationId: oiCreateQuickPromotion
      tags: [quickPromotions]
      summary: Submit a quick promotion request (no auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OiCreateQuickPromotionBody"
      responses:
        "201":
          description: Quick promotion submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OiQuickPromotion"
  /oi/quick-promotions/{id}:
    patch:
      operationId: oiUpdateQuickPromotion
      tags: [quickPromotions]
      summary: Update quick promotion status (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OiUpdateQuickPromotionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OiQuickPromotion"
components:
  schemas:
...
    Influencer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        niche:
          type: string
        followersCount:
          type: integer
        instagramUrl:
          type: ["string", "null"]
        tiktokUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, location, niche, followersCount, instagramUrl, tiktokUrl, youtubeUrl, phone, whatsappNumber, bio, profileImageUrl, createdAt, updatedAt]
...
    OiLoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    OiAuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/OiUser"
      required: [token, user]
    OiUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, client]
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, fullName, role, companyName, phone, createdAt]
...
    OiCampaign:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, posted, completed, cancelled]
        postUrl:
          type: ["string", "null"]
        postedAt:
          type: ["string", "null"]
        clientId:
          type: integer
        influencerId:
          type: integer
        clientName:
          type: ["string", "null"]
        influencerName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, clientId, influencerId, createdAt, updatedAt]
    OiCampaignDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, posted, completed, cancelled]
        postUrl:
          type: ["string", "null"]
        postedAt:
          type: ["string", "null"]
        clientId:
          type: integer
        influencerId:
          type: integer
        client:
          $ref: "#/components/schemas/OiUser"
        influencer:
          $ref: "#/components/schemas/Influencer"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, clientId, influencerId, createdAt, updatedAt]
    OiCreateCampaignBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        clientId:
          type: integer
        influencerId:
          type: integer
      required: [title, clientId, influencerId]
    OiUpdateCampaignBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, posted, completed, cancelled]
        postUrl:
          type: ["string", "null"]
...
    OiQuickPromotion:
      type: object
      properties:
        id:
          type: integer
        influencerId:
          type: integer
        influencerName:
          type: ["string", "null"]
        contactName:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        description:
          type: string
        promoType:
          type: string
        status:
          type: string
          enum: [new, contacted, in_progress, done]
        createdAt:
          type: string
      required: [id, influencerId, contactName, description, promoType, status, createdAt]
    OiCreateQuickPromotionBody:
      type: object
      properties:
        influencerId:
          type: integer
        contactName:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        description:
          type: string
        promoType:
          type: string
      required: [influencerId, contactName, description, promoType]
    OiUpdateQuickPromotionBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, in_progress, done]