# lib/api-zod/src/generated/types/oiCreateCampaignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified API for GambiaInfluencers and Our Influencers platforms
 * OpenAPI spec version: 0.2.0
 */

export interface OiCreateCampaignBody {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  budget?: string | null;
  clientId: number;
  influencerId: number;
}