# lib/api-zod/src/generated/types/oiQuickPromotion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified API for GambiaInfluencers and Our Influencers platforms
 * OpenAPI spec version: 0.2.0
 */
import type { OiQuickPromotionStatus } from "./oiQuickPromotionStatus";

export interface OiQuickPromotion {
  id: number;
  influencerId: number;
  /** @nullable */
  influencerName?: string | null;
  contactName: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  description: string;
  promoType: string;
  status: OiQuickPromotionStatus;
  createdAt: string;
}