# lib/api-zod/src/generated/types/oiUpdateCampaignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified API for GambiaInfluencers and Our Influencers platforms
 * OpenAPI spec version: 0.2.0
 */
import type { OiUpdateCampaignBodyStatus } from "./oiUpdateCampaignBodyStatus";

export interface OiUpdateCampaignBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  budget?: string | null;
  status?: OiUpdateCampaignBodyStatus;
  /** @nullable */
  postUrl?: string | null;
}