# lib/api-zod/src/generated/types/oiUpdateQuickPromotionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified API for GambiaInfluencers and Our Influencers platforms
 * OpenAPI spec version: 0.2.0
 */
import type { OiUpdateQuickPromotionBodyStatus } from "./oiUpdateQuickPromotionBodyStatus";

export interface OiUpdateQuickPromotionBody {
  status?: OiUpdateQuickPromotionBodyStatus;
}