# lib/api-zod/src/generated/types/oiUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified API for GambiaInfluencers and Our Influencers platforms
 * OpenAPI spec version: 0.2.0
 */
import type { OiUserRole } from "./oiUserRole";

export interface OiUser {
  id: number;
  email: string;
  fullName: string;
  role: OiUserRole;
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  phone: string | null;
  createdAt: string;
}